Move swing to utility
