+++ src/main/java/oop_lecture/utility/HomeScreen.java
@@ -1,4 +1,4 @@
-package views;
+package oop_lecture.utility;
 
 import javafx.application.Platform;
 import javafx.embed.swing.JFXPanel;
@@ -9,13 +9,13 @@ import javafx.scene.Scene;
 import javax.swing.*;
 import java.io.IOException;
 
-public class InfoScreen extends JFrame {
+public class HomeScreen extends JFrame {
 
-    public InfoScreen() {
+    public HomeScreen() {
         JFXPanel fxPanel = new JFXPanel();
         this.add(fxPanel);
 
-        this.setTitle("Information Page");
+        this.setTitle("Home Page");
         this.setVisible(true);
         this.setSize(1024, 768);
 
@@ -23,8 +23,8 @@ public class InfoScreen extends JFrame {
             @Override
             public void run() {
                 try {
-                    FXMLLoader loader = new FXMLLoader(getClass().getResource("/views/info.fxml"));
-                    loader.setController(new InfoScreenController());
+                    FXMLLoader loader = new FXMLLoader(getClass().getResource("/views/home.fxml"));
+                    loader.setController(new HomeScreenController());
 
                     Parent root = loader.load();
                     Scene scene = new Scene(root);
@@ -37,6 +37,8 @@ public class InfoScreen extends JFrame {
         });
     }
     public static void main(String[] args){
-        new InfoScreen();
+        //Setup home
+
+        new HomeScreen();
     }
 }
