+++ src/main/java/oop_lecture/utility/InfoScreen.java
@@ -1,4 +1,4 @@
-package views;
+package oop_lecture.utility;
 
 import javafx.application.Platform;
 import javafx.embed.swing.JFXPanel;
@@ -9,13 +9,13 @@ import javafx.scene.Scene;
 import javax.swing.*;
 import java.io.IOException;
 
-public class HomeScreen extends JFrame {
+public class InfoScreen extends JFrame {
 
-    public HomeScreen() {
+    public InfoScreen() {
         JFXPanel fxPanel = new JFXPanel();
         this.add(fxPanel);
 
-        this.setTitle("Home Page");
+        this.setTitle("Information Page");
         this.setVisible(true);
         this.setSize(1024, 768);
 
@@ -23,8 +23,8 @@ public class HomeScreen extends JFrame {
             @Override
             public void run() {
                 try {
-                    FXMLLoader loader = new FXMLLoader(getClass().getResource("/views/home.fxml"));
-                    loader.setController(new HomeScreenController());
+                    FXMLLoader loader = new FXMLLoader(getClass().getResource("/views/info.fxml"));
+                    loader.setController(new InfoScreenController());
 
                     Parent root = loader.load();
                     Scene scene = new Scene(root);
@@ -37,8 +37,6 @@ public class HomeScreen extends JFrame {
         });
     }
     public static void main(String[] args){
-        //Setup home
-
-        new HomeScreen();
+        new InfoScreen();
     }
 }
